+++ RELEASENOTES.md
@@ -32,6 +32,9 @@ Release notes
         ID3 v2.4.
     *   Add `MediaMetadata.mediaType` to denote the type of content or the type
         of folder described by the metadata.
+    *   Add `MediaMetadata.isBrowsable` as a replacement for
+        `MediaMetadata.folderType`. The folder type will be deprecated in the
+        next release.
 *   Remove deprecated symbols:
     *   Remove `DefaultAudioSink` constructors, use `DefaultAudioSink.Builder`
         instead.